+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ### **RRV Enhancements**
 **Improvements**
 - Improvement to the Craftables side panel causing massive lag spikes when the inventory is opened.
+
+**Added**
+- Added a new Mutations recipe type.